Merge mermaidAPI functionality
This commit attempts to add mermaidAPI functionality directly to
remark-mermaid. `mermaid.cli`, on which this previously depended
appears to be completely unmaintained, the result of which is that
many features added to mermaid are impossible to use without
forking `mermaid.cli` and uprgrading it oneself.

I tried my best to follow the conventions I saw already in this
package, but I will say that because this now requires more setup
to run, the number of arguments that need to be passed is more
than I'd like. This is something that could probably be solved
with a few well-placed classes here and there probably, but I
didn't want to introduce classes into an otherwise entirly
procedural repo for the first time.

DEPENDENCIES

This commit removes dependency on `mermaid.cli` and `which`.
It adds dependency on `tmp-promise`, which is used for building
an HTML directory for using the mermaidAPI, and `puppeteer`,
which is used for rendering the SVGs themselves.

I also add dependency on `mermaid` itself for including the JS
library for rendering, and 2 fontawesome repos to replicate
the function of `mermaid.cli`.

USE OF ASYNC

I use async heavily. This is something of an inherited style
from `puppeteer`, but it should work well for what we're doing
beneficially, this means that all SVGs are actually rendered
in parallel.

However, this asynchrony also invevitably propagated to caller
functions in index.js. I've tried to keep refactoring index.js
to a minimum.

RENDERING SVGs

This approach is the same as the one used by `mermaid.cli`,
with a few improvements. `mermaid.cli` relies on a postinstall
process to copy its Javascript and CSS dependencies into a `dist`
folder.

I find this approach a little brittle, when these dependencies, if
at all possible should be `require`d in. Additionally, this project
doesn't have a `dist` folder or a postinstall script, so I didn't
want to add one.

I try to `require.resolve` these dependencies instead, and
construct a temporary directory with soft links to these dependencies.

When mermaid loads, it then loads them effecively as though it were
requiring them directly.

OBJECT DECONSTRUCTION

As previously mentioned, I tried to maintain the procedural style
of the pre-existing code. This resulted in a lot more arguments
than before. I've modified the relevant functions to use object
deconstruction parameters to make this easier to read and write.

MINOR CHANGES

- in replaceUrlWithGraph, I found the jsdoc a little confusing
gramatically. I've changed it to something I think is simpler.

- I found the `isMermaid` function unnecessarily complicated
and removed it.  The operation it performs is just string comparison,
but it being broken out into a function gave me the impression that
it did something more complicated. Additionally, it wasn't consistently
used between functions: some did the check themselves.

- In the code I've touched, I've replaced `if (x) { return y }`
with `if (x) return y` where it is a failure case. Personally,
I find this easier to read.

- I've removed the many constants created from imports of `util`.
I considered using object deconstruction to do the same more
succinctly, but ultimately I found that *not* namespacing
was more misleading, so I've simply removed the aliases, and
added `util.` to usage of its members.

- `visit()` didn't seem like it would play well with `async`, so
I created an adaptor, `collectPromises`, which lets us `await`
on completion of all the functions it calls.

- `visitLink` and `visitImage` were literally copy-and-pastes
of the same code. I've replaced both with `visitAll`, which
has a `nodeType` parameter and modified the transformer
appropriately.

- `render` now renders mermaid code to svg code -- no file used.
`renderToFile` performs the same function as `render` used to.